Refactor PKI classes into their own package
This will support generation of keys on the node.
