+++ upup/pkg/fi/k8sapi/cert_secret.go
@@ -19,7 +19,7 @@ package k8sapi
 import (
 	"fmt"
 	"k8s.io/client-go/pkg/api/v1"
-	"k8s.io/kops/upup/pkg/fi"
+	"k8s.io/kops/pkg/pki"
 )
 
 // KeypairSecret is a wrapper around a k8s Secret object that holds a TLS keypair
@@ -27,8 +27,8 @@ type KeypairSecret struct {
 	Namespace string
 	Name      string
 
-	Certificate *fi.Certificate
-	PrivateKey  *fi.PrivateKey
+	Certificate *pki.Certificate
+	PrivateKey  *pki.PrivateKey
 }
 
 // ParseKeypairSecret parses the secret object, decoding the certificate & private-key, if present
@@ -39,7 +39,7 @@ func ParseKeypairSecret(secret *v1.Secret) (*KeypairSecret, error) {
 
 	certData := secret.Data[v1.TLSCertKey]
 	if certData != nil {
-		cert, err := fi.LoadPEMCertificate(certData)
+		cert, err := pki.LoadPEMCertificate(certData)
 		if err != nil {
 			return nil, fmt.Errorf("error parsing certificate in %s/%s: %q", k.Namespace, k.Name, err)
 		}
@@ -47,7 +47,7 @@ func ParseKeypairSecret(secret *v1.Secret) (*KeypairSecret, error) {
 	}
 	keyData := secret.Data[v1.TLSPrivateKeyKey]
 	if keyData != nil {
-		key, err := fi.ParsePEMPrivateKey(keyData)
+		key, err := pki.ParsePEMPrivateKey(keyData)
 		if err != nil {
 			return nil, fmt.Errorf("error parsing key in %s/%s: %q", k.Namespace, k.Name, err)
 		}
