+++ upup/pkg/fi/nodeup/template_functions.go
@@ -27,6 +27,7 @@ import (
 	api "k8s.io/kops/pkg/apis/kops"
 	"k8s.io/kops/pkg/apis/nodeup"
 	"k8s.io/kops/pkg/flagbuilder"
+	"k8s.io/kops/pkg/pki"
 	"k8s.io/kops/upup/pkg/fi"
 	"k8s.io/kops/upup/pkg/fi/secrets"
 	"k8s.io/kops/util/pkg/vfs"
@@ -122,17 +123,17 @@ func (t *templateFunctions) populate(dest template.FuncMap) {
 }
 
 // CACertificate returns the primary CA certificate for the cluster
-func (t *templateFunctions) CACertificate() (*fi.Certificate, error) {
+func (t *templateFunctions) CACertificate() (*pki.Certificate, error) {
 	return t.keyStore.Cert(fi.CertificateId_CA)
 }
 
 // PrivateKey returns the specified private key
-func (t *templateFunctions) PrivateKey(id string) (*fi.PrivateKey, error) {
+func (t *templateFunctions) PrivateKey(id string) (*pki.PrivateKey, error) {
 	return t.keyStore.PrivateKey(id)
 }
 
 // Certificate returns the specified private key
-func (t *templateFunctions) Certificate(id string) (*fi.Certificate, error) {
+func (t *templateFunctions) Certificate(id string) (*pki.Certificate, error) {
 	return t.keyStore.Cert(id)
 }
